+++ app/src/main/java/com/github/taasonei/randomfox/ui/view/ViewPagerAdapter.kt
@@ -1,11 +1,9 @@
-package com.github.taasonei.randomfox
+package com.github.taasonei.randomfox.ui.view
 
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.FragmentManager
 import androidx.lifecycle.Lifecycle
 import androidx.viewpager2.adapter.FragmentStateAdapter
-import com.github.taasonei.randomfox.ui.FavouritesListFragment
-import com.github.taasonei.randomfox.ui.RecentImageFragment
 
 private const val NUM_TABS = 2
