+++ app/src/main/java/com/github/taasonei/randomfox/ui/viewmodels/RecentImageViewModel.kt
@@ -1,4 +1,4 @@
-package com.github.taasonei.randomfox.viewmodel
+package com.github.taasonei.randomfox.ui.viewmodels
 
 import android.app.Application
 import android.util.Log
@@ -6,10 +6,10 @@ import androidx.lifecycle.AndroidViewModel
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.viewModelScope
-import com.github.taasonei.randomfox.database.DatabaseFox
-import com.github.taasonei.randomfox.model.FoxPhoto
-import com.github.taasonei.randomfox.model.Status
-import com.github.taasonei.randomfox.repository.FoxRepository
+import com.github.taasonei.randomfox.data.database.DatabaseFox
+import com.github.taasonei.randomfox.ui.model.FoxPhoto
+import com.github.taasonei.randomfox.ui.model.Status
+import com.github.taasonei.randomfox.data.repository.FoxRepository
 import kotlinx.coroutines.launch
 import java.io.IOException
